added hero image, sidebar component for sheets, billing sheet, and order sheet update and separation
added hero image, sidebar component for sheets, billing sheet, and order sheet update and separation
